Factorize Dart_CObject
